Fix permission issue while moving files to trash
Users needed to provide username, uid and gid by changing the constants
in the script; so, it can be run via sudo command. The script should be
running with administrative privileges in order to move old apps to trash; brew
installing process will be forked and run as a child process without
administrative privileges.